test calculator import refactor.
