+++ mdsuite/calculators/__init__.py
@@ -28,13 +28,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, List
 
-# isort: off
-from .calculator import Calculator
-from .trajectory_calculator import TrajectoryCalculator
-
-# isort: on
-
 from .angular_distribution_function import AngularDistributionFunction
+from .calculator import Calculator
 from .coordination_number_calculation import CoordinationNumbers
 from .einstein_diffusion_coefficients import EinsteinDiffusionCoefficients
 from .einstein_distinct_diffusion_coefficients import (
