+++ mdsuite/calculators/einstein_helfand_ionic_conductivity.py
@@ -32,8 +32,8 @@ import numpy as np
 import tensorflow as tf
 from tqdm import tqdm
 
-from mdsuite.calculators import TrajectoryCalculator
 from mdsuite.calculators.calculator import call
+from mdsuite.calculators.trajectory_calculator import TrajectoryCalculator
 from mdsuite.database import simulation_properties
 from mdsuite.utils.calculator_helper_methods import fit_einstein_curve
 from mdsuite.utils.units import boltzmann_constant, elementary_charge
