+++ mdsuite/calculators/green_kubo_ionic_conductivity.py
@@ -37,8 +37,8 @@ import tensorflow_probability as tfp
 from bokeh.models import Span
 from tqdm import tqdm
 
-from mdsuite.calculators import TrajectoryCalculator
 from mdsuite.calculators.calculator import call
+from mdsuite.calculators.trajectory_calculator import TrajectoryCalculator
 from mdsuite.database import simulation_properties
 from mdsuite.utils.units import boltzmann_constant, elementary_charge
