+++ mdsuite/calculators/green_kubo_self_diffusion_coefficients.py
@@ -38,8 +38,8 @@ from bokeh.plotting import figure
 from scipy.integrate import cumtrapz
 from tqdm import tqdm
 
-from mdsuite.calculators import TrajectoryCalculator
 from mdsuite.calculators.calculator import call
+from mdsuite.calculators.trajectory_calculator import TrajectoryCalculator
 from mdsuite.database import simulation_properties
